move iterator impl out into tests to bake
