+++ strum/src/lib.rs
@@ -272,93 +272,6 @@ impl<T: EnumRepr> OpaqueRepr<T> {
     }
 }
 
-impl<
-        R: num_traits::PrimInt
-            + core::ops::BitOrAssign
-            + num_traits::WrappingShr
-            + num_traits::WrappingShl,
-        E: EnumRepr<EnumT = E, Repr = R>,
-        O: Copy + EnumRepr<EnumT = E, Repr = R>,
-    > EnumMaskIter for O
-{
-    type I = EnumMaskIterator<R, E, O>;
-
-    fn mask_iter(&self) -> EnumMaskIterator<R, E, O> {
-        EnumMaskIterator {
-            mask: self.to_repr(),
-            shift: 0,
-            phantom: core::marker::PhantomData,
-        }
-    }
-}
-
-pub trait EnumMaskIter: Sized
-where
-    Self: EnumRepr,
-    Self::Repr: core::ops::BitOr + core::ops::BitOrAssign,
-    <Self as EnumRepr>::EnumT: EnumRepr,
-{
-    type I: Iterator<Item = Self::EnumT>;
-
-    fn mask_iter(&self) -> Self::I;
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct EnumMaskIterator<
-    R: num_traits::int::PrimInt
-        + std::ops::BitOrAssign
-        + num_traits::ops::wrapping::WrappingShl
-        + num_traits::ops::wrapping::WrappingShr,
-    T: EnumRepr<Repr = R, EnumT = T>,
-    O: EnumRepr<Repr = R, EnumT = T>,
-> {
-    mask: R,
-    shift: u32,
-    phantom: core::marker::PhantomData<(O, R)>,
-}
-
-impl<
-        R: num_traits::WrappingShr
-            + num_traits::WrappingShl
-            + num_traits::int::PrimInt
-            + std::ops::BitOrAssign,
-        T: EnumRepr<Repr = R, EnumT = T>,
-        O: EnumRepr<Repr = R, EnumT = T>,
-    > Iterator for EnumMaskIterator<R, T, O>
-{
-    type Item = <T as EnumRepr>::EnumT;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        // This can doubtlessly be improved
-        if self.shift >= std::mem::size_of::<R>() as u32 * 8 {
-            return None;
-        }
-
-        let tz: u32 = self.mask.trailing_zeros();
-        let tz = if tz < std::mem::size_of::<R>() as u32 * 8 {
-            tz
-        } else {
-            std::mem::size_of::<R>() as u32 * 8 - 1
-        };
-        let discr = ((self.mask.wrapping_shr(tz)) & num_traits::identities::one())
-            .wrapping_shl(self.shift + tz);
-
-        let one_u32: u32 = num_traits::identities::one();
-        self.mask = self.mask.wrapping_shr(tz + one_u32);
-        let shift_lhs: u32 = core::ops::Add::<u32>::add(tz, one_u32);
-        self.shift += shift_lhs;
-        T::EnumT::cvt_from_repr(discr)
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        // FIXME This assumes mask bits only contain enum repr items...
-        // And there is only one bit to each enum discriminant.
-        (
-            self.mask.count_ones() as usize,
-            Some(self.mask.count_ones() as usize),
-        )
-    }
-}
 
 // FIXME derive ExactSizeIterator...
